requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-ollama==0.2.1
langchain-experimental==0.3.4
tavily-python>=0.3.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
pypdf